Fix: - raising explicit error when running benchmark without cuda - merging quant, dequant and gemm code into one file - removing depricated int4/int8 comparison
